+++ json_intl/lib/src/json_intl_delegate_builtin.dart
@@ -5,7 +5,6 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
-import 'package:json_intl/json_intl.dart';
 
 import 'json_intl.dart';
 import 'json_intl_data.dart';